% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalDantzig.R
\name{plot.causalDantzig}
\alias{plot.causalDantzig}
\title{Plot coefficients paths from a "causalDantzig" object with regularization}
\usage{
\method{plot}{causalDantzig}(x, ...)
}
\arguments{
\item{x}{fitted "causalDantzig" model with regularization}

\item{...}{Additional inputs to generic plot function (not used).}
}
\description{
Gives a plot of the coefficient paths for a fitted "causalDantzig" object with regularization.
}
\seealso{
\code{causalDantzig} and \code{summary} method.
}
