% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalDantzig.R
\name{summary.causalDantzig}
\alias{summary.causalDantzig}
\title{Summarizes "causalDantzig" fits}
\usage{
\method{summary}{causalDantzig}(object, ...)
}
\arguments{
\item{object}{fitted "causalDantzig" model with or without regularization}

\item{...}{Additional inputs to generic summary function (not used).}
}
\description{
\code{summary} method for class "causalDantzig".
}
\seealso{
\code{causalDantzig} and \code{plot} method.
}
