\name{summary.InvariantCausalPrediction}
\alias{summary.InvariantCausalPrediction}
\title{
Summary of "InvariantCausalPrediction" objects
}
\description{
Prints confidence intervals for invariant causal prediction (output of \code{\link{ICP}} function)
}
\usage{
\method{summary}{InvariantCausalPrediction}(object, maxshow = 50, ...)
}
\arguments{
  \item{object}{
Object of class "InvariantCausalPrediction", as generated by function \code{\link{ICP}}.
}
  \item{maxshow}{
Maximal number of variables to show in the summary.
}
  \item{\dots}{
Additional inputs to generic summary function (not used).
}
}
\value{
Does not return an object.
}
\references{
Jonas Peters, Peter Buhlmann, Nicolai Meinshausen (2015):

 Causal inference using invariant prediction: identification and confidence intervals

 arxiv preprint \url{http://arxiv.org/abs/1501.01332}
}
\author{
Nicolai Meinshausen <meinshausen@stat.math.ethz.ch>
}
\seealso{
\code{\link{ICP}}
}

\keyword{Causality }
\keyword{Regression }
\keyword{Classification }
