\name{InvariantCausalPrediction-package}
\alias{InvariantCausalPrediction-package}
\alias{InvariantCausalPrediction}
\docType{package}
\title{
Invariant causal prediction
}
\description{
Confidence intervals for causal prediction in a regression setting. An experimental version is also available for classification. 
}
\details{
\tabular{ll}{
Package: \tab InvariantCausalPrediction\cr
Type: \tab Package\cr
Version: \tab 0.6-1\cr
Date: \tab 2016-05-02\cr
License: \tab GPL\cr
}
Confidence intervals can be computed with function \code{ICP}, both for regression and binary classification. \code{print}, \code{plot} and \code{summary} methods are available. 
}
\author{
Nicolai Meinshausen <meinshausen@stat.math.ethz.ch>

Maintainer: Nicolai Meinshausen <meinshausen@stat.math.ethz.ch>
}
